% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_data.R
\name{design_data}
\alias{design_data}
\title{use scDesign to simulate scRNA-seq data}
\usage{
design_data(realcount, S = 1e+08, ncell, ngroup = 1, pUp = 0.05,
  pDown = 0.05, fU = 5, fL = 1.5, ncores = 1)
}
\arguments{
\item{realcount}{A numeric matrix with rows representing genes and columns representing cells. 
Gene names are given as row names.}

\item{S}{A number specifying the total number of RNA-seq reads. Default to 1e8.}

\item{ncell}{An integer specifying the number of cells. When \code{ngroup > 1},
\code{ncell} is the number of cells in each cell state.}

\item{ngroup}{An integer giving the number of cell states to simulate. Defaults to 1.}

\item{pUp}{A value between 0 and 1 specifying the proportion of up regulated genes
between two adjacent cell states. Defaults to 0.05 and only used when \code{ngroup > 1}.}

\item{pDown}{A value between 0 and 1 specifying the proportion of down regulated genes
between two adjacent cell states. Defaults to 0.05 and only used when \code{ngroup > 1}.}

\item{fU}{A value specifying the upper bound of fold changes of differentially expressed genes.
Deaults to 5.}

\item{fL}{A value specifying the lower bound of fold changes of differentially expressed genes.
Deaults to 1.5.}

\item{ncores}{An integer specifying the number of cores used for parallel computation.
Defaults to 1.}
}
\value{
When \code{ngroup = 1}, \code{design_data} returns a simulated count matrix with
rows representing genes and columns representing cells.
When \code{ngroup > 1}, \code{design_data} returns a list of \code{ngroup} elements. 
The g-th element corresponds to the g-th cell state, and is a list containing three elements:
\describe{
  \item{count:}{a count matrix with rows representing genes and columns representing cells;}
  \item{genesUp:}{a character vector giving the names of up-regulated genes from state g-1 to g;}
  \item{genesDown:}{a character vector giving the names of down-regulated genes from state g-1 to g.}
}
}
\description{
use scDesign to simulate scRNA-seq data
}
\author{
Wei Vivian Li, \email{liw@ucla.edu}

Jingyi Jessica Li, \email{jli@stat.ucla.edu}
}
