% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_data.R
\name{design_joint}
\alias{design_joint}
\title{use scDesign to make experimental design assuming two cell states are sequenced together}
\usage{
design_joint(realcount1, realcount2, prop1, prop2, S = 1e+08,
  ncell = round(2^seq(6, 13, 1)), B = 100, de_method = "ttest",
  p_thre = 10^seq(-2, -6, -1), plot_dir = "./", ncores = 1, rank = 1000)
}
\arguments{
\item{realcount1}{A numeric matrix with rows representing genes and columns representing cells (cell state 1). 
Gene names are given as row names.}

\item{realcount2}{A numeric matrix with rows representing genes and columns representing cells (cell state 2). 
Gene names are given as row names.}

\item{prop1}{A number giving the proportion of state 1 cells in the cell population.}

\item{prop2}{A number giving the proportion of state 2 cells in the cell population.}

\item{S}{A number specifying the total number of RNA-seq reads for the cell population. Default to 1e8.}

\item{ncell}{An integer vector specifying the total number of cells to sequence. 
Defaults to \code{2^seq(6,13,1)}.}

\item{B}{An integer giving the number of experiments to repeat in order the calculate 
the average DE analysis accuracy. Defaults to 100.}

\item{de_method}{A character specifying the differential expression analysis method to use.
Currently supports "ttest" (default) or "mast".}

\item{p_thre}{A numeric vector specifying the FDR thresholds used to identify 
differentially expressed genes. Defaults to \code{10^seq(-2,-6,-1)}.}

\item{plot_dir}{A character giving the directory to save experimental design results
Defaults to "./".}

\item{ncores}{An integer specifying the number of cores used for parallel computation.
Defaults to 1.}

\item{rank}{An integer specifying the number of top DE genes to identify 
from scImpute's results as the standard in DE analysis. Defaults to 1000.}
}
\value{
A list of five elments:
\describe{
  \item{precision:}{a matrix of precision. }
  \item{recall:}{a matrix of recall (true positive rate).}
  \item{TN:}{a matrix of TN (true negative rate).}
  \item{F1:}{a matrix of F1 (precision vs. recall).}
  \item{F2:}{a matrix of F2 (TN vs. recall).}
}
In all the matrices, rows correspond to different FDR thresholds and 
columns correspond to the cell numbers  specified in \code{ncell}.
\code{design_joint} also writes the list to design_summary.txt and
saves it to \code{plot_dir}. 
The corresponding plots are also saved to \code{plot_dir}.
}
\description{
use scDesign to make experimental design assuming two cell states are sequenced together
}
\author{
Wei Vivian Li, \email{liw@ucla.edu}

Jingyi Jessica Li, \email{jli@stat.ucla.edu}
}
